Move feature-repo tests to a separate file, improve error logging
